fasthtml_playground/pyproject.toml
[project]
name = "fasthtml-playground"
version = "0.1.0"
description = ""
authors = [
    {name = "Andreas Daiminger",email = "andreas@chaosisnotrandomitisrhythmic.xyz"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "python-fasthtml (>=0.12.6,<0.13.0)",
    "loguru (>=0.7.3,<0.8.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
isort = "^6.0.1"
black = "^25.1.0"


[tool.poetry.group.server.dependencies]
boto3 = "^1.37.23"
opensearch-py = "^2.8.0"
fastapi = "^0.115.12"

